+++ app/src/main/java/org/simple/clinic/patient/onlinelookup/SecondsDuration.kt
@@ -1,0 +1,8 @@
+package org.simple.clinic.patient.onlinelookup
+
+import com.squareup.moshi.JsonQualifier
+import kotlin.annotation.Retention
+
+@Retention(AnnotationRetention.RUNTIME)
+@JsonQualifier
+annotation class SecondsDuration